# storage/api-docs/api-docs-user.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My College  API swagger documentation",
        "description": "My College User API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "besanmarwan2000@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/user/family/members/permission-types": {
            "get": {
                "tags": [
                    "Family Permissions"
                ],
                "summary": "Get family permission types",
                "description": "Retrieve all available permission types that can be assigned to a family member.",
                "operationId": "getFamilyPermissionTypes",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/language"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Family permission types retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Family permission types retrieved successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/FamilyPermissionType"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated."
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/v1/user/family/members/{patientFamilyMember}/permissions": {
            "put": {
                "tags": [
                    "Family Permissions"
                ],
                "summary": "Update family member permissions",
                "description": "Update the permissions of a family member for the authenticated patient.",
                "operationId": "updateFamilyMemberPermissions",
                "parameters": [
                    {
                        "name": "patientFamilyMember",
                        "in": "path",
                        "description": "Patient family member ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 7
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "permissions"
                                ],
                                "properties": {
                                    "permissions": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "enum": [
                                                "view_medications",
                                                "view_dialysis_sessions",
                                                "view_fluid_data",
                                                "view_symptoms",
                                                "view_labs",
                                                "view_measurements",
                                                "receive_alerts"
                                            ]
                                        },
                                        "example": [
                                            "view_medications",
                                            "view_fluid_data",
                                            "receive_alerts"
                                        ],
                                        "minItems": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Family member permissions updated successfully."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Only the patient can manage family member permissions."
                    },
                    "404": {
                        "description": "Family member relationship not found."
                    },
                    "422": {
                        "description": "Validation error."
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/v1/user/family/members/{patientFamilyMember}/revoke": {
            "patch": {
                "tags": [
                    "Family Permissions"
                ],
                "summary": "Revoke family member access",
                "description": "Revoke a family member's access to the authenticated patient's data.",
                "operationId": "revokeFamilyMemberAccess",
                "parameters": [
                    {
                        "name": "patientFamilyMember",
                        "in": "path",
                        "description": "Patient family member ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 7
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Family member access revoked successfully."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Family member relationship not found."
                    },
                    "422": {
                        "description": "Family member access is already revoked."
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/v1/user/family/members/{patientFamilyMember}/restore": {
            "patch": {
                "tags": [
                    "Family Permissions"
                ],
                "summary": "Restore family member access",
                "description": "Restore a revoked family member's access to the authenticated patient's data and assign new permissions.",
                "operationId": "restoreFamilyMemberAccess",
                "parameters": [
                    {
                        "name": "patientFamilyMember",
                        "in": "path",
                        "description": "Patient family member ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 7
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "permissions"
                                ],
                                "properties": {
                                    "permissions": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "enum": [
                                                "view_medications",
                                                "view_dialysis_sessions",
                                                "view_fluid_data",
                                                "view_symptoms",
                                                "view_labs",
                                                "view_measurements",
                                                "receive_alerts"
                                            ]
                                        },
                                        "example": [
                                            "view_medications",
                                            "view_fluid_data",
                                            "receive_alerts"
                                        ],
                                        "minItems": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Family member access restored successfully."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "404": {
                        "description": "Family member relationship not found."
                    },
                    "422": {
                        "description": "Family member access is not revoked or validation failed."
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "FamilyPermissionType": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 12
                    },
                    "name": {
                        "type": "string",
                        "example": "view_medicinne"
                    },
                    "label": {
                        "type": "string",
                        "example": "مشاهدة سجل الادوية"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
            "language": {
                "name": "language",
                "in": "header",
                "description": "response language : ar for arabic || en for english",
                "required": false,
                "schema": {
                    "type": "string"
                },
                "examples": {
                    "ar": {
                        "summary": "Arabic",
                        "value": "ar"
                    },
                    "en": {
                        "summary": "English",
                        "value": "en"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "send Bearer token",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    },
...
    "security": [
        {
            "api_key": []
        }
    ]
}